moved Login actions to new file called LoginActions.js. removed actions from index.js
